Restructure repo: config.py to config/config.py + tests
